feat(physic): seed is working
