+++ src/webgl/controllers/Loaders/sources.ts
@@ -37,21 +37,6 @@ const Sources: ISource[] = [
     path: "/models/earthv5.glb",
   },
   {
-    name: "housev1model",
-    type: "gltfModel",
-    path: "/models/housev9.glb",
-  },
-  {
-    name: "grassv1model",
-    type: "gltfModel",
-    path: "/models/grassv9.glb",
-  },
-  {
-    name: "housev1texture",
-    type: "texture",
-    path: "/textures/bakedHousev9.jpg",
-  },
-  {
     name: "seedMatCap",
     type: "texture",
     path: "/textures/seedMatCap.jpg",
@@ -188,6 +173,66 @@ const Sources: ISource[] = [
     type: "texture",
     path: "/textures/matcap.jpeg",
   },
+  // HOUSE GRASS
+  {
+    name: "grassv1model",
+    type: "gltfModel",
+    path: "/models/HouseScene/grass.glb",
+  },
+  {
+    name: "housev1texture",
+    type: "texture",
+    path: "/textures/HouseScene/bakedHouse.jpg",
+  },
+  // HOUSE SCENE
+  {
+    name: "housev1model",
+    type: "gltfModel",
+    path: "/models/HouseScene/house.glb",
+  },
+  {
+    name: "housev1texture",
+    type: "texture",
+    path: "/textures/HouseScene/bakedHouse.jpg",
+  },
+  // CURSOR
+  {
+    name: "cursor-texture",
+    type: "texture",
+    path: "/textures/cursor.png",
+  },
+  // BILLY
+  {
+    name: "billy-model",
+    type: "gltfModel",
+    path: "/models/Characters/Billy/billyModel.glb",
+  },
+  {
+    name: "billy-model-test",
+    type: "gltfModel",
+    path: "/models/character.glb",
+  },
+  {
+    name: "billy-texture",
+    type: "texture",
+    path: "/textures/Characters/Billy/billyTextureV2.jpg",
+  },
+  // TREES
+  {
+    name: "big-tree-model",
+    type: "gltfModel",
+    path: "/models/Trees/bigTree.glb",
+  },
+  {
+    name: "medium-tree-model",
+    type: "gltfModel",
+    path: "/models/Trees/mediumTree.glb",
+  },
+  {
+    name: "small-tree-model",
+    type: "gltfModel",
+    path: "/models/Trees/smallTree.glb",
+  },
   // SEED
   {
     name: "seed-model",
@@ -198,12 +243,6 @@ const Sources: ISource[] = [
     name: "poppingtrees-seed-texture",
     type: "texture",
     path: "/textures/poppingtrees-seed.jpg",
-  },
-  // CURSOR
-  {
-    name: "cursor-texture",
-    type: "texture",
-    path: "/textures/cursor.png",
   },
 ];
