+++ src/webgl/world/homeScene/Billy/Shaders/vertex.glsl
@@ -1,0 +1,7 @@
+varying vec2 vUv;
+
+void main() {
+    gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
+
+    vUv = uv;
+}
