+++ src/webgl/world/homeScene/SeedGame/Lumberjack/LumberjackSettings.ts
@@ -7,7 +7,7 @@ const lamberjackSettings = {
   visualizeDepth: 10,
   gravity: -250,
   playerSpeed: 10,
-  physicsSteps: 5,
+  physicsSteps: 10,
   basePosition: new Vector3(0, 0.1, 1),
 };
 export default lamberjackSettings;
